maps_bot/requirements.txt
playwright
g4f
pandas
openpyxl
python-dotenv
colorama
customtkinter
geopy
